chore: move files and remove unnecessary code
