+++ .eslintignore
@@ -1,4 +1,3 @@
-examples/TesterApp
 node_modules
 website/node_modules
 website/vendor
