+++ examples/TesterApp/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: '@callstack/eslint-config',
+};
