--- examples/TesterApp/error.ts
@@ -1,3 +1,0 @@
-export function raise() {
-  throw new Error('test_error');
-}
