+++ examples/TesterApp/index.js
@@ -1,11 +1,7 @@
-/**
- * @format
- */
-
 import { AppRegistry } from 'react-native';
-import App from './App';
-import { name as appName } from './app.json';
 import { ChunkManager, Chunk } from '../../client';
+import App from './src/App';
+import { name as appName } from './app.json';
 
 ChunkManager.configureResolver(async (chunkId) => {
   if (__DEV__) {
