+++ examples/TesterApp/src/App.js
@@ -1,0 +1,45 @@
+import React from 'react';
+import {
+  SafeAreaView,
+  ScrollView,
+  StatusBar,
+  useColorScheme,
+  View,
+} from 'react-native';
+import { Colors, Header } from 'react-native/Libraries/NewAppScreen';
+
+import { RemoteContainer } from './RemoteContainer';
+import { AsyncContainer } from './AsyncContainer';
+import { Section } from './Section';
+
+const App = () => {
+  const isDarkMode = useColorScheme() === 'dark';
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+  return (
+    <SafeAreaView style={backgroundStyle}>
+      <StatusBar barStyle={isDarkMode ? 'light-content' : 'dark-content'} />
+      <ScrollView
+        contentInsetAdjustmentBehavior="automatic"
+        style={backgroundStyle}
+      >
+        <Header />
+        <View
+          style={{
+            backgroundColor: isDarkMode ? Colors.black : Colors.white,
+          }}
+        >
+          <Section title="Async chunk">
+            <AsyncContainer />
+          </Section>
+          <Section title="Remote chunks">
+            <RemoteContainer />
+          </Section>
+        </View>
+      </ScrollView>
+    </SafeAreaView>
+  );
+};
+
+export default App;
