+++ examples/TesterApp/src/Async.js
@@ -1,9 +1,6 @@
 import React from 'react';
 import { Text } from 'react-native';
 
-
 export default function Async() {
-  return (
-    <Text>Async</Text>
-  );
+  return <Text>Async: this text comes from async chunk</Text>;
 }
