+++ examples/TesterApp/src/Remote.js
@@ -2,5 +2,5 @@ import React from 'react';
 import { Text } from 'react-native';
 
 export default function Remote() {
-  return <Text>Remote</Text>;
+  return <Text>Remote: this text comes from remote chunk</Text>;
 }
